Moved classification_options.h to c/task/processor
