+++ tensorflow_lite_support/c/task/processor/BUILD
@@ -15,3 +15,8 @@ cc_library(
     name = "bounding_box",
     hdrs = ["bounding_box.h"],
 )
+
+cc_library(
+    name = "classification_options",
+    hdrs = ["classification_options.h"],
+)
